Extract send message as a component
